+++ src/main/java/com/h3/reservation/calendar/CalendarController.java
@@ -1,4 +1,4 @@
-package com.h3.reservation;
+package com.h3.reservation.calendar;
 
 import com.google.api.client.util.DateTime;
 import com.google.api.services.calendar.Calendar;
