+++ src/main/java/com/h3/reservation/calendar/WebConfig.java
@@ -1,4 +1,4 @@
-package com.h3.reservation;
+package com.h3.reservation.calendar;
 
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
